requirements.txt
black==25.1.0
flake8==7.2.0
isort==6.0.1
bandit==1.7.5
mypy==1.4.1
pre-commit
pytest
futu-api
fastapi
uvicorn
pydantic
numpy
pandas